Added file to view the source of any example.  This will be more useful later, when there are lots more examples.
Updated documentation to look better, fix typos, fix examples, make use of
viewSource.php, link to SourceForge, etc.

Updated functions.inc to have a cleaner looking layout (in my opinion) and
it will be much easier to exapand without having to worry if it will fit on
the screen or not.